[SimpleInvDyn] Add waist and contact6D tasks to CoM and Posture ones
Add outputs (q_des, v_des, u) and enum on output control modes to use the entity in torque or velocity.
q_des and v_des are the integration of the desired acceleration computed by TSID
u is the result of the desired torque computed + PD on posture des.
Works with talos-torque-control package for sinusoid traj on the CoM in velocity control.