[Inv_Dyn] Add function actFrame to transform the CoM and Feet frames
This transformation is necessary to have the tasks acting in the same frames as the trajectory reference ones. If they are expressed in the same frame (i.e. the world) the transformation is equal to the identity matrix.  
I need this to use the trajectories of the multicontact-locomotion-planning framework, for the walks on stairs and platforms which have a reference frame different from the world.